pyproject.toml
[project]
name = "core-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.12.14",
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "mistralai>=1.9.3",
    "mysql-connector>=2.2.9",
    "openai>=1.97.1",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "redis>=6.2.0",
    "tenacity>=9.1.2",
    "tqdm>=4.67.1",
    "uvicorn>=0.35.0",
]
